Move color to css mapper.
